# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AI Gateway API
  version: 0.1.0
  description: |
    Multi-tenant AI gateway. Centralizes provider key isolation, routing,
    fallback, rate limiting, and usage observability. The AI Capability
    Node of the Microservice Node Governance Platform.
  contact:
    name: AI Platform Team
    email: lead.ai-gateway@example.com
servers:
  - url: http://localhost:8080
    description: Local single-container deployment
  - url: https://ai-gateway.internal.example.com
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /v1/me:
    get:
      tags: [me]
      operationId: getMe
      summary: Current user profile.
      responses:
        '200':
          description: Profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/me/tokens:
    get:
      tags: [me]
      operationId: listMyTokens
      summary: List personal access tokens. Raw token values are never returned here.
      responses:
        '200':
          description: List of tokens (prefix only).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PatDto' }
    post:
      tags: [me]
      operationId: createMyToken
      summary: Create a personal access token. Raw token is returned ONCE.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePatRequest' }
      responses:
        '200':
          description: Token created. The 'token' field is shown only here.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatePatResponse' }
  /v1/me/tokens/{id}:
    delete:
      tags: [me]
      operationId: revokeMyToken
      summary: Revoke a personal access token.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '204': { description: Revoked. }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/me/models:
    get:
      tags: [me]
      operationId: getMyAvailableModels
      summary: List models available to the user based on active partner keys in their account.
      responses:
        '200':
          description: List of available models.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AvailableModelDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or PAT (prefix `aigw_`)
      description: |
        Pass either a JWT obtained from `/v1/auth/login` or a Personal
        Access Token (prefix `aigw_`) in the `Authorization: Bearer <token>`
        header. The server auto-dispatches between the two schemes based on
        the prefix.
...
  responses:
...
    Unauthorized:
      description: Authentication required or failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }

    # ─── Auth ───
...
    UserDto:
      type: object
      properties:
        id:          { type: integer, format: int64 }
        email:       { type: string }
        role:        { type: string, enum: [user, admin] }
        status:      { type: string, enum: [active, disabled] }
        displayName: { type: string, nullable: true }
        createdAt:   { type: string, format: date-time }

    # ─── PATs ───
    CreatePatRequest:
      type: object
      required: [name]
      properties:
        name:          { type: string, minLength: 1, maxLength: 255 }
        expiresInDays: { type: integer, minimum: 0, maximum: 3650, nullable: true }
        useCaveman:
          type: boolean
          default: false
          description: When true, responses generated with this PAT are automatically compressed using Caveman style.
    CreatePatResponse:
      type: object
      required: [id, name, token, createdAt, useCaveman, responseStyle]
      properties:
        id:        { type: integer, format: int64 }
        name:      { type: string }
        token:
          type: string
          description: Raw token. Returned ONLY at creation. Format `aigw_<24 url-safe chars>`.
          example: aigw_5K3xQs8nLp_vT7yU2eHmA1Bw
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        useCaveman: { type: boolean }
        responseStyle: { type: string, enum: [normal, caveman] }
    PatDto:
      type: object
      properties:
        id:          { type: integer, format: int64 }
        name:        { type: string }
        tokenPrefix: { type: string, example: aigw_5K3xQs8n }
        lastUsedAt:  { type: string, format: date-time, nullable: true }
        expiresAt:   { type: string, format: date-time, nullable: true }
        createdAt:   { type: string, format: date-time }
        useCaveman:  { type: boolean }
        responseStyle: { type: string, enum: [normal, caveman] }

    # ─── Account keys ───
...
    AvailableModelDto:
      type: object
      required: [code, name, partnerCode, partnerName]
      properties:
        code:               { type: string }
        name:               { type: string }
        partnerCode:        { type: string }
        partnerName:        { type: string }
        defaultTemperature: { type: number }
        defaultMaxTokens:   { type: integer }